-: Warning fixed
git-svn-id: http://cpp-events.googlecode.com/svn/trunk@95 67599820-88c8-11de-9c77-2f49170b69f2

+++ tests/unit/Test_EqualMemberAddress/TestClass.cpp
@@ -4,7 +4,7 @@ static int magic = 0x11223344;
 
 int TestClass::testOutlineFunction()
 {
-	return someData_ * reinterpret_cast<int>(&magic);
+	return someData_ * (int)reinterpret_cast<size_t>(&magic);
 }
 
 int TestClass::testOutlineFunctionWithStaticVariable()
